Java: Add two cache classes
Implement several caches that can store key-value pairs and create soft
references to the values. The caches can optionally hold a limited
number of key-value pairs and are thread safe.

The implementation in the `cache` directory makes use of the
`synchronized` keyword, and the implementation in the
`concurrent-read-cache` directory makes use of a read-write lock.  While
I haven't done performance benchmarking, I expect the former
implementation to make more intelligent caching decisions, and I expect
the latter implementation to have higher performance in read-heavy
multi-threaded applications.

The heavy use of the `synchronized` keyword on the former implementation
may seem like overkill. however, its' necessary. From the `HashMap`
documentation:

     If multiple threads access a hash map concurrently, and at least
     one of the threads modifies the map structurally, it must be
     synchronized externally.  (A structural modification is any
     operation that adds or deletes one or more mappings; merely
     changing the value associated with a key that an instance already
     contains is not a structural modification.)

Depending on the exact use case, users of both types of cache objects
might still need to use the `synchronized` and `volatile` keywords.  The
former is useful in situations where a read from the cache affects a
subsequent write, and the latter is useful any time one thread reads
from the cache, a second thread writes to the cache, and the first
thread reads from the cache again.