Properly cache dependencies
Currently, the environment requires a fair amount of set-up before the
average-rating Hadoop application can be run. For example,
`~/data/hadoop-2.9.1/` must exist. In addition, the process of
installing `data/ml-latest-small.zip` can be quite inefficient: if it
doesn't exist, it's always downloaded from the internet. This file can
easily be blown away with a `git clean -dfx`, and re-downloading it
isn't always an option.

Improve the process of satisfying dependencies. If a dependency doesn't
exist in `data/`, then copy it from `~/.cache/average-rating/`, and if
that file doesn't exist, then download it into the cache.